track previousCommand string to avoid dupes
This has the potential to introduce bugs.
We are basically saying if the current commandString is the same
as the first one, don't send it across the socket, its probably
a duplicate. but its not necessarily a duplicate. We should unset
this.previousCommandString when theres a chance its going to fall out
of sync.